pyproject.toml
[project]
name = "study-learning-microsoft-foundry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent-framework-core>=1.0.0rc5",
    "azure-ai-agents>=1.1.0",
    "azure-ai-documentintelligence>=1.0.2",
    "azure-ai-projects>=2.0.1",
    "azure-ai-textanalytics>=5.4.0",
    "azure-identity>=1.25.3",
    "azure-keyvault-secrets>=4.10.0",
    "ipykernel>=7.2.0",
    "openai>=2.28.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
]
